backend/requirements.txt
# MathNote Backend Dependencies

# ==================== Core ====================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# ==================== AWS ====================
boto3>=1.34.0
botocore>=1.34.0

# ==================== AI / LLM ====================
openai>=1.10.0
requests>=2.31.0
google-generativeai>=0.3.0
aiohttp>=3.9.0
nvidia-riva-client

# ==================== Video / Audio Processing ====================
opencv-python-headless>=4.9.0
ffmpeg-python>=0.2.0
openai-whisper>=20231117
# faster-whisper>=0.10.0  # Alternative for faster inference

# ==================== Image Processing ====================
Pillow>=10.2.0
scikit-image>=0.22.0  # For SSIM calculation

# ==================== Utilities ====================
python-dotenv>=1.0.0
httpx>=0.26.0

# ==================== Development ====================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0

# ==================== Optional (YouTube Download) ====================
yt-dlp>=2024.1.0
